pyproject.toml
[project]
name = "cartola-analytics"
version = "0.1.0"
description = "Cartola FC data pipelines and analytics"
authors = [
    {name = "Cadu Santos", email = "carlos.silva150995@gmail.com"}
]
license = {text = "GPL-3.0-or-later"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.13.1"
black = "^25.9.0"
mypy = "^1.18.2"
pytest = "^8.4.2"
respx = "^0.22.0"
types-pyyaml = "^6.0.12.20250915"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src"]
extend-exclude = ["data", "docs"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
mypy_path = "src"
files = ["src/cartola_analytics"]
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
strict_optional = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["src/tests"]



[tool.ruff.lint.isort]
known-first-party = ["cartola_analytics"]



[tool.poetry.scripts]
cartola-fetch = "cartola_analytics.cli:main"
